refactor: dto 디렉토리 생성
