+++ src/main/java/com/kuit/conet/auth/apple/AppleUserProvider.java
@@ -3,7 +3,7 @@ package com.kuit.conet.auth.apple;
 import com.kuit.conet.utils.JwtParser;
 import com.kuit.conet.utils.auth.PublicKeyGenerator;
 import com.kuit.conet.common.exception.InvalidTokenException;
-import com.kuit.conet.dto.response.ApplePlatformUserResponse;
+import com.kuit.conet.dto.response.auth.ApplePlatformUserResponse;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
