+++ src/main/java/com/kuit/conet/dto/response/auth/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.kuit.conet.dto.response;
+package com.kuit.conet.dto.response.auth;
 
 import lombok.*;
