+++ src/main/java/com/kuit/conet/service/AuthService.java
@@ -9,12 +9,12 @@ import com.kuit.conet.common.exception.UserException;
 import com.kuit.conet.dao.UserDao;
 import com.kuit.conet.domain.Platform;
 import com.kuit.conet.domain.User;
-import com.kuit.conet.dto.request.LoginRequest;
-import com.kuit.conet.dto.request.PutOptionTermAndNameRequest;
-import com.kuit.conet.dto.response.AgreeTermAndPutNameResponse;
-import com.kuit.conet.dto.response.ApplePlatformUserResponse;
-import com.kuit.conet.dto.response.KakaoPlatformUserResponse;
-import com.kuit.conet.dto.response.LoginResponse;
+import com.kuit.conet.dto.request.auth.LoginRequest;
+import com.kuit.conet.dto.request.auth.PutOptionTermAndNameRequest;
+import com.kuit.conet.dto.response.auth.AgreeTermAndPutNameResponse;
+import com.kuit.conet.dto.response.auth.ApplePlatformUserResponse;
+import com.kuit.conet.dto.response.auth.KakaoPlatformUserResponse;
+import com.kuit.conet.dto.response.auth.LoginResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
